+++ Frends.As2.SendMessage/Frends.As2.SendMessage.Tests/docker/config/config.xml
@@ -206,7 +206,7 @@
            defaults="$properties.pollerConfigBase.defaults$"
            sendfilename="$properties.pollerConfigBase.sendfilename$"
            mimetype="$properties.pollerConfigBase.mimetype$"
-           process_files_in_paralllel="$properties.pollerConfigBase.process_files_in_paralllel$"
+           process_files_in_parallel="$properties.pollerConfigBase.process_files_in_paralllel$"
            max_parallel_files="$properties.pollerConfigBase.max_parallel_files$"/>
    <partnerships classname="org.openas2.partner.XMLPartnershipFactory"
                  filename="$properties.partnership_file$"
