+++ Frends.As2.SendMessage/Frends.As2.SendMessage/Definitions/Connection.cs
@@ -66,5 +66,6 @@ public class Connection
     /// </summary>
     /// <example>application/zip</example>
     [DisplayFormat(DataFormatString = "Text")]
+    [DefaultValue("text/plain")]
     public string ContentTypeHeader { get; set; }
 }
