+++ Frends.As2.SendMessage/Frends.As2.SendMessage/Definitions/Error.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace Frends.As2.SendMessage.Definitions;
 
 /// <summary>
@@ -15,5 +17,5 @@ public class Error
     /// Additional information about the error.
     /// </summary>
     /// <example>object { Exception Exception }</example>
-    public dynamic AdditionalInfo { get; set; }
+    public Exception AdditionalInfo { get; set; }
 }
