+++ Frends.As2.SendMessage/Frends.As2.SendMessage/Frends.As2.SendMessage.csproj
@@ -18,7 +18,7 @@
 
     <ItemGroup>
         <PackageReference Include="BouncyCastle.Cryptography" Version="2.6.1" />
-        <PackageReference Include="Frends.NSoftware.Activation" Version="1.1.0" />
+<!--        <PackageReference Include="Frends.NSoftware.Activation" Version="1.1.0" />-->
         <PackageReference Include="MimeKit" Version="4.13.0" />
         <PackageReference Include="nsoftware.IPWorksEDI" Version="24.0.9373" />
         <PackageReference Include="StyleCop.Analyzers" Version="1.1.118">
@@ -37,6 +37,9 @@
 
     <ItemGroup>
         <None Include="FrendsTaskMetadata.json" Pack="true" PackagePath="/">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+        <None Include="../CHANGELOG.md" Pack="true" PackagePath="/">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
     </ItemGroup>
